fix task-4
